# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SaleItem {
  id: number;
  saleId: number;
  productId: number | null;
  name: string;
  size: string;
  qty: number;
  rate: number;
  amount: number;
}

export interface Sale {
  id: number;
  billNo: string;
  date: string;
  customer: string;
  /** @nullable */
  phone: string | null;
  /** @nullable */
  note: string | null;
  paymentMethod: string;
  total: number;
  createdAt: string;
  items: SaleItem[];
}

export interface CreateSaleItemBody {
  name: string;
  productId: number;
  size: string;
  qty: number;
  rate: number;
}

export interface CreateSaleBody {
  billNo: string;
  date: string;
  customer: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  note?: string | null;
  paymentMethod: string;
  items: CreateSaleItemBody[];
}

export interface SalesSummary {
  total: number;
  count: number;
}

export interface MonthlyTotal {
  month: string;
  total: number;
  count: number;
}

export interface Product {
  id: number;
  name: string;
  category: string;
  size: string;
  price: number;
  stock: number;
  currentStock: number;
  totalStockIn: number;
  totalSold: number;
  current_stock: number;
  total_stock_in: number;
  total_sold: number;
  createdAt: string;
}

export interface CreateProductBody {
  name: string;
  category: string;
  size: string;
  price: number;
  stock: number;
}

export type ListSalesParams = {
  year?: number;
  month?: number;
};

export type GetMonthlyTotalsParams = {
  year?: number;
};