# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserResponse"
        "401":
          description: Authentication required
  /auth/change-password:
    post:
      operationId: changeOwnPassword
      tags: [auth]
      summary: Change current user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordBody"
      responses:
        "200":
          description: Password changed
        "400":
          description: Validation error
        "401":
          description: Authentication required
...
  /products/{id}/stock-in:
    post:
      operationId: stockInProduct
      tags: [products]
      summary: Add stock to a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockInBody"
      responses:
        "201":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}/stock-movements:
    get:
      operationId: listProductStockMovements
      tags: [products]
      summary: List stock movements for one product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stock movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockMovement"
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Validation error
  /products/{id}:
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
components:
  schemas:
...
    AuthUserResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/CurrentUser"
      required:
        - user
    CurrentUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, user]
        active:
          type: boolean
      required:
        - id
        - email
        - name
        - role
        - active
    ChangePasswordBody:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
        confirmPassword:
          type: string
          minLength: 8
      required:
        - currentPassword
        - newPassword
        - confirmPassword
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        size:
          type: string
        price:
          type: number
        stock:
          type: integer
        currentStock:
          type: integer
        totalStockIn:
          type: integer
        totalSold:
          type: integer
        current_stock:
          type: integer
        total_stock_in:
          type: integer
        total_sold:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - category
        - size
        - price
        - stock
        - currentStock
        - totalStockIn
        - totalSold
        - current_stock
        - total_stock_in
        - total_sold
        - createdAt
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        size:
          type: string
        price:
          type: number
        stock:
          type: integer
      required:
        - name
        - category
        - size
        - price
        - stock
    StockInBody:
      type: object
      properties:
        quantity:
          type: integer
        note:
          type: ["string", "null"]
      required:
        - quantity
    StockMovement:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        type:
          type: string
          enum: [IN, OUT]
        quantity:
          type: integer
        reason:
          type: string
          enum: [purchase, sale, adjustment, return, manual]
        saleId:
          type: ["integer", "null"]
        createdBy:
          type: ["integer", "null"]
        createdByEmail:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - productId
        - productName
        - type
        - quantity
        - reason
        - saleId
        - createdBy
        - createdByEmail
        - note
        - createdAt