# lib/api-zod/src/generated/types/createProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductBody {
  name: string;
  category: string;
  size: string;
  price: number;
  stock: number;
}