# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  category: string;
  size: string;
  price: number;
  stock: number;
  currentStock: number;
  totalStockIn: number;
  totalSold: number;
  current_stock: number;
  total_stock_in: number;
  total_sold: number;
  createdAt: string;
}